Обновление счетчиков навигации при клике в карточке
