+++ src/model/navigation.js
@@ -1,10 +1,43 @@
 import Observer from "../utils/observer.js";
 import {CATEGORIES} from "../const.js";
+import {LANG} from "../lang";
+import {UPDATE_TYPE} from "../const";
 
 export default class NavigationModel extends Observer {
   constructor() {
     super();
     this._activeCategory = CATEGORIES.ALL;
+  }
+
+  getCounters() {
+    const navigationMap = {
+      [CATEGORIES.ALL]: {
+        itemTitle: LANG.ALL_MOVIES,
+        countEntries(films) {
+          return films.filter((film) => film).length;
+        }
+      },
+      [CATEGORIES.WATCHLIST]: {
+        itemTitle: LANG.WATCHLIST_CAP,
+        countEntries(films) {
+          return films.filter((film) => film.isInWatchlist).length;
+        }
+      },
+      [CATEGORIES.HISTORY]: {
+        itemTitle: LANG.HISTORY_CAP,
+        countEntries(films) {
+          return films.filter((film) => film.isInHistory).length;
+        }
+      },
+      [CATEGORIES.FAVORITES]: {
+        itemTitle: LANG.FAVORITES_CAP,
+        countEntries(films) {
+          return films.filter((film) => film.isInFavorites).length;
+        }
+      }
+    };
+
+    return navigationMap;
   }
 
   setFilter(updateType, category) {
@@ -14,5 +47,9 @@ export default class NavigationModel extends Observer {
 
   getFilter() {
     return this._activeCategory;
+  }
+
+  updateCounters() {
+    this._notify(UPDATE_TYPE.MINOR)
   }
 }
